Remove alias of compose and pipe
|> is well known in F#, OCaml and SML as the pipe operator, but the
signature [1] is different than what we have in hymn [2], to avoid
confusion, it should be removed before it is too late.  By doing so,
<| as an alias of compose should also be removed.  I did not choose .
(dot) for compose in the first place, as it is already been used in hy.
The unicode symbol ∘ is a better choice for compose, and it was the
symbol I used before public release of hymn, I removed it because LaTeX
does not support unicode characters out-of-box, I did not want to use
some extra LaTeX unicode packages and custom preamble just for the sole
purpose of compiling documentation into pdf file.

[1] in F#, OCaml and SML, the signature of |> is
    (|>) : 'a -> ('a -> 'b) -> 'b

[2] in hymn, the signature of compose (thus <| as well), in ML's term,
    is
    ('a -> 'b) -> ('b -> 'c) -> 'a -> 'c