Removed function apply and using #* and #** instead
Resort to what @MelomanCool implemented (using `#*` and `#**` directly
instead of `apply`) in the draft of the pull request.

I was thinking a defmacro version of apply, until I figured out the way
I structured the macros will cause circular import (require) in this
case, `hymn.macros` requires macro `|` from `hymn.types.either` which in
turn requires `hymn.macros.apply`, the logically place the macro `apply`
should be. MelomanCool's implementation in first place was the right
way, so be it.