Removed problematic macro import and made examples work
We used to be able to import (require) macros from other modules in one
module and those imported (required) macros will be visible to the
namespace that imports (requires) this very module, such as how it works
with import statements in python.

For example:

mod_a.py:
def fa():
    pass

mod_b.py:
from a import fa

In python, fa can be used as mod_b.fa as well. hy macros used to work
like that, it no longer does.